+++ src/integrations/mod.rs
@@ -1,6 +1,9 @@
 #[cfg(feature = "bytemuck")]
 mod bytemuck;
 
+#[cfg(feature = "fixed")]
+mod fixed;
+
 #[cfg(feature = "mint")]
 mod mint;